% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodpar.R
\name{methodpar}
\alias{methodpar}
\title{A methopar Function}
\usage{
methodpar(Rf = 0, gap = 12, p = 0.95)
}
\arguments{
\item{Rf}{risk free rate, in same period as you returns. Default to 0.}

\item{gap}{numeric number of periods from start of series to use to train risk calculation. Default to 12.}

\item{p}{confidence level for calculation. Default to 0.95.}
}
\description{
This function allows you to set parameters for corresponding plotmethod in plots()
}
\examples{
methodpar(Rf, gap, p)
}
\keyword{methodpar}

