% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{A plots Function}
\usage{
plots(xts, plotmethod = "default", methodpar)
}
\arguments{
\item{xts}{xts Object.}

\item{plotmethod}{Mention your plotmethod from supported package. At present, only charts.PerformanceSummary is working. Default to plot.xts method.}

\item{methodpar}{mention your methodpar function}
}
\description{
This function allows you to plot according to chosen plotmethod from various packages
}
\examples{
plots(xts, plotmethod = "charts.PerformanceSummary", methodpar)
}
\keyword{plots}

